+++ packages/resource/test/index.test.ts
@@ -1,14 +1,5 @@
-import { createEffect, createResource, createSignal, on } from "solid-js";
+import { catchError, createEffect, createResource, createRoot, createSignal, on } from "solid-js";
 import { describe, test, expect, vi, afterAll, beforeEach, beforeAll } from "vitest";
-import { testEffect } from "@solidjs/testing-library";
-
-class AbortError extends Error {
-  constructor(msg: string) {
-    super(msg);
-  }
-  name = "AbortError";
-}
-
 import {
   makeAbortable,
   createAggregated,
@@ -17,6 +8,33 @@ import {
   makeRetrying,
   createDeepSignal,
 } from "../src/index.js";
+
+export function testEffect<T extends any = void>(
+  fn: (done: (result: T) => void) => void,
+): Promise<T> {
+  let done: (result: T) => void;
+  let fail: (error: any) => void;
+  let promise = new Promise<T>((resolve, reject) => {
+    done = resolve;
+    fail = reject;
+  });
+  createRoot(dispose => {
+    catchError(() => {
+      fn(result => {
+        done(result);
+        dispose();
+      });
+    }, fail);
+  });
+  return promise;
+}
+
+class AbortError extends Error {
+  constructor(msg: string) {
+    super(msg);
+  }
+  name = "AbortError";
+}
 
 beforeAll(() => {
   vi.useFakeTimers();