+++ packages/storage/src/cookies.ts
@@ -1,12 +1,13 @@
 import { getRequestEvent, isServer, type RequestEvent } from "solid-js/web";
-import { StorageProps, StorageSignalProps, StorageWithOptions } from "./types.js";
-import { addClearMethod, addWithOptionsMethod } from "./tools.js";
-import { createStorage, createStorageSignal } from "./storage.js";
+import { SyncStorageWithOptions } from "./index.js";
+import { addWithOptionsMethod, addClearMethod } from "./tools.js";
 
-export type CookieOptions = CookieProperties & {
-  getRequestHeaders?: () => Headers;
-  getResponseHeaders?: () => Headers;
-};
+export type CookieOptions =
+  | (CookieProperties & {
+      getRequestHeaders?: () => Headers;
+      getResponseHeaders?: () => Headers;
+    })
+  | undefined;
 
 type CookiePropertyTypes = {
   domain?: string;
@@ -18,7 +19,9 @@ type CookiePropertyTypes = {
   sameSite?: "None" | "Lax" | "Strict";
 };
 
-type CookieProperties = { [key in keyof CookiePropertyTypes]: CookiePropertyTypes[key] };
+type CookieProperties = {
+  [key in keyof CookiePropertyTypes]: CookiePropertyTypes[key];
+};
 
 const cookiePropertyMap = {
   domain: "Domain",
@@ -34,7 +37,7 @@ function serializeCookieOptions(options?: CookieOptions) {
   if (!options) return "";
   const result = Object.entries(options)
     .map(([key, value]) => {
-      const serializedKey = cookiePropertyMap[key as keyof CookiePropertyTypes];
+      const serializedKey: string | undefined = cookiePropertyMap[key as keyof CookiePropertyTypes];
       if (!serializedKey) return undefined;
 
       if (value instanceof Date) return `${serializedKey}=${value.toUTCString()}`;
@@ -47,16 +50,17 @@ function serializeCookieOptions(options?: CookieOptions) {
 }
 
 function deserializeCookieOptions(cookie: string, key: string) {
-  return cookie.match(`(^|;)\\s*${key}\\s*=\\s*([^;]+)`)?.pop() ?? null;
+  const found = cookie.match(`(^|;)\\s*${key}\\s*=\\s*([^;]+)`)?.pop();
+  return found != null ? decodeURIComponent(found) : null;
 }
 
 const getRequestHeaders = isServer
-  ? () => getRequestEvent()?.request?.headers || new Headers()
+  ? () => getRequestEvent()?.request.headers || new Headers()
   : () => new Headers();
 const getResponseHeaders = isServer
   ? () =>
-      (getRequestEvent() as RequestEvent & { response: Response })?.response?.headers ||
-      new Headers()
+      (getRequestEvent() as (RequestEvent & { response: Response }) | undefined)?.response
+        .headers || new Headers()
   : () => new Headers();
 
 /**
@@ -78,7 +82,7 @@ const getResponseHeaders = isServer
  * ```
  * Also, you can use its _read and _write properties to change reading and writing
  */
-export const cookieStorage: StorageWithOptions<CookieOptions> = addWithOptionsMethod(
+export const cookieStorage: SyncStorageWithOptions<CookieOptions> = addWithOptionsMethod(
   addClearMethod({
     _read: isServer
       ? (options?: CookieOptions) => {
@@ -102,9 +106,9 @@ export const cookieStorage: StorageWithOptions<CookieOptions> = addWithOptionsMe
           const responseHeaders = getResponseHeaders();
           const currentCookies =
             responseHeaders
-              ?.get("Set-Cookie")
+              .get("Set-Cookie")
               ?.split(", ")
-              ?.filter(cookie => cookie && !cookie.startsWith(`${key}=`)) ?? [];
+              .filter(cookie => cookie && !cookie.startsWith(`${key}=`)) ?? [];
           responseHeaders.set(
             "Set-Cookie",
             [...currentCookies, `${key}=${value}${serializeCookieOptions(options)}`].join(", "),
@@ -116,10 +120,17 @@ export const cookieStorage: StorageWithOptions<CookieOptions> = addWithOptionsMe
     getItem: (key: string, options?: CookieOptions) =>
       deserializeCookieOptions(cookieStorage._read(options), key),
     setItem: (key: string, value: string, options?: CookieOptions) => {
-      cookieStorage._write(key, value, options);
+      cookieStorage._write(
+        key,
+        value.replace(/[\u00c0-\uffff\&;]/g, c => encodeURIComponent(c)),
+        options,
+      );
     },
     removeItem: (key: string, options?: CookieOptions) => {
-      cookieStorage._write(key, "deleted", { ...options, expires: new Date(0) });
+      cookieStorage._write(key, "deleted", {
+        ...options,
+        expires: new Date(0),
+      });
     },
     key: (index: number, options?: CookieOptions) => {
       let key: string | null = null;
@@ -147,25 +158,3 @@ export const cookieStorage: StorageWithOptions<CookieOptions> = addWithOptionsMe
     },
   }),
 );
-
-/**
- * creates a reactive store but bound to document.cookie
- * @deprecated in favor of makePersisted
- */
-export const createCookieStorage = <T, O = CookieOptions, A = StorageWithOptions<CookieOptions>>(
-  props?: Omit<StorageProps<T, A, O>, "api">,
-) => createStorage<O, T>({ ...props, api: cookieStorage } as any);
-
-/**
- * creates a reactive signal, but bound to document.cookie
- * @deprecated in favor of makePersisted
- */
-export const createCookieStorageSignal = <
-  T,
-  O = CookieOptions,
-  A = StorageWithOptions<CookieOptions>,
->(
-  key: string,
-  initialValue?: T,
-  props?: Omit<StorageSignalProps<T, A, O>, "api">,
-) => createStorageSignal<T, O>(key, initialValue, { ...props, api: cookieStorage } as any);