+++ packages/list/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @solid-primitives/list
 
+## 0.0.101
+
+### Patch Changes
+
+- 1ea2a09: Use isDev const from solid-js/web instead if a string
+
 ## 0.0.100
 
 Initial release of the package.