+++ packages/resource/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @solid-primitives/resource
 
+## 0.3.1
+
+### Patch Changes
+
+- 89117a6: storage: expose init promise/value, resource: docs clarification
+
+## 0.3.0
+
+### Minor Changes
+
+- d7d0d70: resource: new primitive createAbortable (makeAbortable + cleanup)
+
 ## 0.2.0
 
 ### Minor Changes